+++ backend/Cell.py
@@ -7,6 +7,13 @@ class Color(Enum):
     WHITE = 2
     VOID = -1
 
+    @classmethod
+    def swap_turn(self, curr):
+        if curr == Color.BLACK:
+            return Color.WHITE
+        elif curr == Color.WHITE:
+            return Color.BLACK
+
 
 class Cell:
     def __init__(self):
